requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0
jinja2>=3.1.2
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
httpx>=0.25.0
